Move function to utils, and actually call it 🤣 Chapeau, TypeScript!
