+++ src/renderer/renderScene.ts
@@ -47,8 +47,10 @@ import {
   TransformHandles,
   TransformHandleType,
 } from "../element/transformHandles";
-import { viewportCoordsToSceneCoords } from "../utils";
+import { viewportCoordsToSceneCoords, supportsEmoji } from "../utils";
 import { UserIdleState } from "../excalidraw-app/collab/types";
+
+const hasEmojiSupport = supportsEmoji();
 
 const strokeRectWithRotation = (
   context: CanvasRenderingContext2D,
@@ -172,19 +174,6 @@ const renderLinearPointHandles = (
   context.lineWidth = lineWidth;
   context.translate(-sceneState.scrollX, -sceneState.scrollY);
   context.strokeStyle = origStrokeStyle;
-};
-
-// From https://github.com/Modernizr/Modernizr/blob/master/feature-detects/emoji.js
-const supportsEmoji = () => {
-  const canvas = document.createElement("canvas");
-  const ctx = canvas.getContext("2d");
-  const backingStoreRatio = ctx.backingStorePixelRatio || 1;
-  const offset = 12 * backingStoreRatio;
-  ctx.fillStyle = "#f00";
-  ctx.textBaseline = "top";
-  ctx.font = "32px Arial";
-  ctx.fillText("\ud83d\udc28", 0, 0); // U+1F428 KOALA
-  return ctx.getImageData(offset, offset, 1, 1).data[0] !== 0;
 };
 
 export const renderScene = (
@@ -495,7 +484,7 @@ export const renderScene = (
 
     const username = sceneState.remotePointerUsernames[clientId];
     let usernameAndIdleState;
-    if (supportsEmoji) {
+    if (hasEmojiSupport) {
       usernameAndIdleState = `${username ? `${username} ` : ""}${
         userState === UserIdleState.AWAY
           ? "⚫️"
